oci: Add OCI-SIF docs. General restructure.
To try and mention OCI-SIF in a sensible manner, I've restructured the
OCI page considerably.

The intent is that it now gives a reasonable brief overview of what
OCI-SIF is, and where that sits.. i.e.

Fixes #183

Not intended to be a full solution for the other related issues,
though due to the restructuring / rewriting it will have partly
addressed them.